# docs/api-docs.json
  {
    "openapi": "3.1.0",
    "info": {
      "title": "OpenAPI definition",
      "version": "v0"
    },
    "servers": [
      {
        "url": "http://localhost:8082",
        "description": "Generated server url"
      }
    ],
    "paths": {
      "/auth/validate": {
        "post": {
          "tags": [
            "auth-controller"
          ],
          "operationId": "validate",
          "requestBody": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidateTokenRequest"
                }
              }
            },
            "required": true
          },
          "responses": {
            "200": {
              "description": "OK",
              "content": {
                "*/*": {
                  "schema": {
                    "$ref": "#/components/schemas/ValidateTokenResponse"
                  }
                }
              }
            }
          }
        }
      },
      "/auth/register": {
        "post": {
          "tags": [
            "auth-controller"
          ],
          "operationId": "register",
          "requestBody": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegisterRequest"
                }
              }
            },
            "required": true
          },
          "responses": {
            "200": {
              "description": "OK",
              "content": {
                "*/*": {
                  "schema": {
                    "$ref": "#/components/schemas/UserResponse"
                  }
                }
              }
            }
          }
        }
      },
      "/auth/refresh": {
        "post": {
          "tags": [
            "auth-controller"
          ],
          "operationId": "refresh",
          "requestBody": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RefreshTokenRequest"
                }
              }
            },
            "required": true
          },
          "responses": {
            "200": {
              "description": "OK",
              "content": {
                "*/*": {
                  "schema": {
                    "$ref": "#/components/schemas/AuthResponse"
                  }
                }
              }
            }
          }
        }
      },
      "/auth/login": {
        "post": {
          "tags": [
            "auth-controller"
          ],
          "operationId": "login",
          "requestBody": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginRequest"
                }
              }
            },
            "required": true
          },
          "responses": {
            "200": {
              "description": "OK",
              "content": {
                "*/*": {
                  "schema": {
                    "$ref": "#/components/schemas/AuthResponse"
                  }
                }
              }
            }
          }
        }
      },
      "/users": {
        "get": {
          "tags": [
            "user-controller"
          ],
          "operationId": "getAllUsers",
          "responses": {
            "200": {
              "description": "OK",
              "content": {
                "*/*": {
                  "schema": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/UserResponse"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "/users/{id}": {
        "get": {
          "tags": [
            "user-controller"
          ],
          "operationId": "getUserById",
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "integer",
                "format": "int64"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "OK",
              "content": {
                "*/*": {
                  "schema": {
                    "$ref": "#/components/schemas/UserResponse"
                  }
                }
              }
            }
          }
        }
      },
      "/users/me": {
        "get": {
          "tags": [
            "user-controller"
          ],
          "operationId": "me",
          "responses": {
            "200": {
              "description": "OK",
              "content": {
                "*/*": {
                  "schema": {
                    "$ref": "#/components/schemas/UserResponse"
                  }
                }
              }
            }
          }
        }
      }
    },
    "components": {
      "schemas": {
        "ValidateTokenRequest": {
          "type": "object",
          "properties": {
            "token": {
              "type": "string",
              "minLength": 1
            }
          },
          "required": [
            "token"
          ]
        },
        "ValidateTokenResponse": {
          "type": "object",
          "properties": {
            "valid": {
              "type": "boolean"
            },
            "username": {
              "type": "string"
            },
            "userId": {
              "type": "integer",
              "format": "int64"
            },
            "roles": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        },
        "RegisterRequest": {
          "type": "object",
          "properties": {
            "username": {
              "type": "string",
              "maxLength": 255,
              "minLength": 3
            },
            "password": {
              "type": "string",
              "maxLength": 255,
              "minLength": 6
            }
          },
          "required": [
            "password",
            "username"
          ]
        },
        "UserResponse": {
          "type": "object",
          "properties": {
            "id": {
              "type": "integer",
              "format": "int64"
            },
            "username": {
              "type": "string"
            },
            "enabled": {
              "type": "boolean"
            }
          }
        },
        "RefreshTokenRequest": {
          "type": "object",
          "properties": {
            "refreshToken": {
              "type": "string",
              "minLength": 1
            }
          },
          "required": [
            "refreshToken"
          ]
        },
        "AuthResponse": {
          "type": "object",
          "properties": {
            "accessToken": {
              "type": "string"
            },
            "refreshToken": {
              "type": "string"
            }
          }
        },
        "LoginRequest": {
          "type": "object",
          "properties": {
            "username": {
              "type": "string",
              "minLength": 1
            },
            "password": {
              "type": "string",
              "minLength": 1
            }
          },
          "required": [
            "password",
            "username"
          ]
        }
      }
    }
  }