# src/main/java/com/innowise/Authentication_Service/dto/AuthResponse.java
package com.innowise.Authentication_Service.dto;

import lombok.Builder;
import lombok.Data;

@Data
@Builder
public class AuthResponse {
    private String accessToken;
    private String refreshToken;
}